modify: split contract impl file.
