+++ contract/Contracts.HamsterWoodsContract/HamsterWoodsContractConstants.cs
@@ -4,7 +4,7 @@ public class HamsterWoodsContractConstants
 {
     public const string BeanPassSymbol = "BEANPASS-1";
     public const string HalloweenBeanPassSymbol = "BEANPASS-2";
-    public const string BeanSymbol = "BEAN";
+    public const string AcornSymbol = "ACORNS";
 
     public const int DailyMaxPlayCount = 20;
     public const int DailyPlayCountResetHours = 0;
